+++ ext/singularity/deployer.go
@@ -19,10 +19,10 @@ import (
 // c.f. https://github.com/HubSpot/Singularity/blob/master/Docs/reference/configuration.md#limits
 
 // Singularity DeployID must be <50
-const maxDeployIDLen = 50
+const maxDeployIDLen = 49
 
 // Singularity RequestID must be <100
-const maxRequestIDLen = 100
+const maxRequestIDLen = 99
 
 // maxVersionLen needs to account for the separator character
 // between the version string and the UUID string.
@@ -265,10 +265,11 @@ func MakeRequestID(depID sous.DeploymentID) (string, error) {
 	digest := depID.Digest()
 
 	reqBase := fmt.Sprintf("%s-%s-%s-%s-%x", sn, dd, fl, cl, digest)
-	if len(reqBase) < maxRequestIDLen {
-		return reqBase, nil
+
+	if len(reqBase) > maxRequestIDLen {
+		return reqBase[:(maxRequestIDLen)], nil
 	}
-	return reqBase[:(maxRequestIDLen - 1)], nil
+	return reqBase, nil
 }
 
 func computeDeployID(d *sous.Deployable) string {
@@ -288,9 +289,8 @@ func computeDeployID(d *sous.Deployable) string {
 		uuidEntire,
 	}, "_")
 
-	if len(depBase) <= maxDeployIDLen {
-		return depBase
+	if len(depBase) > maxDeployIDLen {
+		return depBase[:(maxDeployIDLen)]
 	}
-
-	return depBase[:(maxDeployIDLen - 1)]
+	return depBase
 }